move types into .d.ts file and import them into the src from there
